feat: 修改sql表中exec_count的数据类型，重命名test包

+++ easy-cron-scheduler/src/main/resources/data/schema.sql
@@ -20,8 +20,8 @@ CREATE TABLE IF NOT EXISTS ecs_exec_detail
     init_time        DATETIME     NOT NULL,
     end_time         DATETIME              DEFAULT NULL,
     last_exec_time   DATETIME              DEFAULT NULL,
-    exec_count       INT          NOT NULL DEFAULT 0,
-    max_exec_count   INT          NOT NULL DEFAULT 0,
+    exec_count       BIGINT       NOT NULL DEFAULT 0,
+    max_exec_count   BIGINT       NOT NULL DEFAULT 0,
     state            INT          NOT NULL DEFAULT 0,
     created_at       DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at       DATETIME     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
